deploy-tool: get rid of own_manifest_base_url
Terraform output included both `own_manifest_base_url` and
`manifest_bucket`, both of which referred to the same storage. The
former was used to check for the existence of manifests via HTTP GET;
the latter to upload new manifests to a cloud storage bucket. This
commit eliminates `own_manifest_base_url`. To that end,
`manifest/writer.go` is renamed to `manifest/storage.go`, and we
introduce interfaces that capture both uploading and fetching manifests.

Closes #751